// frontend/src/api/models/UpdateServiceDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateServiceDto = {
    /**
     * Tên loại phí
     */
    name?: string;
    /**
     * Tháng (YYYY-MM)
     */
    month?: string;
    /**
     * Số tiền của loại phí này
     */
    totalAmount?: number;
    /**
     * Trạng thái: paid hoặc unpaid
     */
    status?: string;
};